storage: move node_api into its own header file
This is a pure refactor commit. The storage node abstraction is going to
become a larger thing and can use its own header.
