+++ src/v/cluster/node/local_monitor.cc
@@ -16,6 +16,7 @@
 #include "config/configuration.h"
 #include "config/node_config.h"
 #include "storage/api.h"
+#include "storage/node.h"
 #include "storage/types.h"
 #include "utils/human.h"
 #include "vassert.h"
