+++ src/v/cluster/node/local_monitor.h
@@ -12,6 +12,7 @@
 #include "cluster/node/types.h"
 #include "config/property.h"
 #include "resource_mgmt/storage.h"
+#include "storage/node.h"
 #include "storage/types.h"
 #include "units.h"
